separated SpaceGroup, SpaceGroupIrreps, SymmetryOperation into 3 files. SpaceGrouyp.from_cell - instead of __init__ SpaceGroupBare unified with SpaceGroup
