+++ irrep/cli.py
@@ -394,7 +394,7 @@ def cli(
                     verbosity, 1)
         log_message("Warning: -kpnames not specified. Only traces of "
                     "symmetry operations will be calculated. Remember that "
-                    "-kpnames must be specified to identify irreps", 
+                    "-kpnames must be specified to identify irreps",
                     verbosity, 1)
 
     # parse input arguments into lists if supplied
